Start fixing the test-toolkit
TODO: Make modules for Spigot and Paper
