+++ commandapi-platforms/commandapi-bukkit/commandapi-bukkit-core/src/main/java/dev/jorel/commandapi/BukkitExecutable.java
@@ -463,8 +463,9 @@ extends BukkitExecutable<Impl>
 	 * @return this command builder
 	 */
 	default Impl executesFeedbackForwarding(FeedbackForwardingCommandExecutor executor) {
-		if (CommandAPIBukkit.platform().activePlatform() != Platform.PAPER) {
-			throw new PlatformException("Attempted to use a FeedbackForwardingCommandExecutor on a non-paper platform (" + CommandAPIBukkit.platform().activePlatform().name() + ")!");
+		Platform platform = CommandAPIBukkit.get().activePlatform();
+		if (platform != Platform.PAPER) {
+			throw new PlatformException("Attempted to use a FeedbackForwardingCommandExecutor on a non-paper platform (" + platform.name() + ")!");
 		}
 		getExecutor().addNormalExecutor(executor);
 		return instance();
@@ -477,8 +478,9 @@ extends BukkitExecutable<Impl>
 	 * @return this command builder
 	 */
 	default Impl executesFeedbackForwarding(FeedbackForwardingExecutionInfo info) {
-		if (CommandAPIBukkit.platform().activePlatform() != Platform.PAPER) {
-			throw new PlatformException("Attempted to use a FeedbackForwardingExecutionInfo on a non-paper platform (" + CommandAPIBukkit.platform().activePlatform().name() + ")!");
+		Platform platform = CommandAPIBukkit.get().activePlatform();
+		if (platform != Platform.PAPER) {
+			throw new PlatformException("Attempted to use a FeedbackForwardingCommandExecutor on a non-paper platform (" + platform.name() + ")!");
 		}
 		getExecutor().addNormalExecutor(info);
 		return instance();
@@ -491,8 +493,9 @@ extends BukkitExecutable<Impl>
 	 * @return this command builder
 	 */
 	default Impl executesFeedbackForwarding(FeedbackForwardingResultingCommandExecutor executor) {
-		if (CommandAPIBukkit.platform().activePlatform() != Platform.PAPER) {
-			throw new PlatformException("Attempted to use a FeedbackForwardingResultingCommandExecutor on a non-paper platform (" + CommandAPIBukkit.platform().activePlatform().name() + ")!");
+		Platform platform = CommandAPIBukkit.get().activePlatform();
+		if (platform != Platform.PAPER) {
+			throw new PlatformException("Attempted to use a FeedbackForwardingCommandExecutor on a non-paper platform (" + platform.name() + ")!");
 		}
 		getExecutor().addResultingExecutor(executor);
 		return instance();
@@ -505,8 +508,9 @@ extends BukkitExecutable<Impl>
 	 * @return this command builder
 	 */
 	default Impl executesFeedbackForwarding(FeedbackForwardingResultingExecutionInfo info) {
-		if (CommandAPIBukkit.platform().activePlatform() != Platform.PAPER) {
-			throw new PlatformException("Attempted to use a FeedbackForwardingResultingExecutionInfo on a non-paper platform (" + CommandAPIBukkit.platform().activePlatform().name() + ")!");
+		Platform platform = CommandAPIBukkit.get().activePlatform();
+		if (platform != Platform.PAPER) {
+			throw new PlatformException("Attempted to use a FeedbackForwardingCommandExecutor on a non-paper platform (" + platform.name() + ")!");
 		}
 		getExecutor().addResultingExecutor(info);
 		return instance();
