+++ commandapi-platforms/commandapi-bukkit/commandapi-bukkit-test-toolkit/pom.xml
@@ -6,7 +6,7 @@
 	<parent>
 		<groupId>dev.jorel</groupId>
 		<artifactId>commandapi-bukkit</artifactId>
-		<version>10.2.0-SNAPSHOT</version>
+		<version>11.0.0-SNAPSHOT</version>
 	</parent>
 
 	<artifactId>commandapi-bukkit-test-toolkit</artifactId>
@@ -45,7 +45,7 @@
 		<!-- Compatibility with CommandAPI Bukkit API -->
 		<dependency>
 			<groupId>dev.jorel</groupId>
-			<artifactId>commandapi-paper-core</artifactId>
+			<artifactId>commandapi-bukkit-core</artifactId>
 			<version>${project.version}</version>
 			<scope>provided</scope>
 		</dependency>
@@ -56,13 +56,6 @@
 			<artifactId>mockito-core</artifactId>
 			<version>5.11.0</version>
 			<scope>compile</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>io.papermc.paper</groupId>
-			<artifactId>paper-api</artifactId>
-			<version>1.20.6-R0.1-SNAPSHOT</version>
-			<scope>provided</scope>
 		</dependency>
 
 		<!-- Pass Brigadier to our dependents -->
