+++ commandapi-platforms/commandapi-bukkit/commandapi-bukkit-test-toolkit/src/test/java/dev/jorel/commandapi/CommandAPIVersionHandler.java
@@ -1,0 +1,14 @@
+package dev.jorel.commandapi;
+
+public abstract class CommandAPIVersionHandler {
+	static LoadContext getPlatform(CommandAPIConfig<?> config) {
+		MockInternalBukkitConfig mockInternalBukkitConfig;
+		if (config instanceof MockCommandAPIBukkitConfig bukkitConfig) {
+			mockInternalBukkitConfig = new MockInternalBukkitConfig(bukkitConfig);
+		} else {
+			throw new IllegalStateException("MockCommandAPIBukkit was loaded with non-MockBukkit config!");
+		}
+
+		return new LoadContext(new MockCommandAPIBukkit(mockInternalBukkitConfig));
+	}
+}
