+++ commandapi-platforms/commandapi-bukkit/commandapi-bukkit-test-toolkit/src/test/java/dev/jorel/commandapi/spying/ExecutionQueueTests.java
@@ -24,7 +24,7 @@ class ExecutionQueueTests extends CommandTestBase {
 	public void setUp() {
 		super.setUp();
 
-		queue = getCommandAPIPlatform().getCommandAPIHandlerSpy().getExecutionQueue();
+		queue = getMockCommandAPI().getCommandAPIHandlerSpy().getExecutionQueue();
 
 		new CommandAPICommand("test")
 			.withOptionalArguments(new BooleanArgument("shouldFail"))
