+++ commandapi-platforms/commandapi-bukkit/pom.xml
@@ -25,8 +25,8 @@
 		<module>commandapi-bukkit-plugin-common</module>
 		<module>commandapi-bukkit-networking-plugin</module>
 
-		<!-- Test toolkit - needs work => TODO
-		<module>commandapi-bukkit-test-toolkit</module> -->
+		<!-- Test toolkit -->
+		<module>commandapi-bukkit-test-toolkit</module>
 
 		<!-- Tests do not compile at this time. This is a big todo
 		<module>commandapi-bukkit-test</module>-->
