+++ commandapi-platforms/commandapi-paper/pom.xml
@@ -25,8 +25,9 @@
 		<module>commandapi-paper-nms</module>
 		<module>commandapi-paper-shade</module>
 		<module>commandapi-paper-plugin</module>
+        <module>commandapi-paper-test-toolkit</module>
 
-		<!-- Do not compile, I only need the code for when I do the documentation -->
+        <!-- Do not compile, I only need the code for when I do the documentation -->
 		<!--<module>commandapi-paper-documentation-code</module> -->
 
 		<!-- Tests do not compile
